Add clippy to CI + address clippy problems (#215)
* Add clippy to CI + address clippy problems

This commit
1. adds `cargo clippy` to .travis.yml
2. addresses clippy concerns and clarifies places where they should be ignored.

* Add clippy component in travis CI

* Fix from_str errors caused by a clippy fix

The issue was that the from_str implementation on Value was moved to the
FromStr trait as recommended by Clippy. This caused many of the tests to
fail since now the FromStr trait must be in scope. Instead of bringing
FromStr into scope everywhere, however, the `Value::from_str(_)` calls
were replaced with `_.parse()` calls as recommended by the std lib
documentation.

* Format crate + rename CI targets for clarity