Fix parsing of borrowed strings
The implementation of `deserialize_str` would incorrectly
call `Visitor::visit_str` when it should have called
`Visitor::visit_borrowed_str`. As per https://serde.rs/lifetimes.html#transient-borrowed-and-owned-data,
`visit_borrowed_str` should be called when the lifetime of the deserialized
string is that of the input.

This should not be a breaking change, because it only adds more guarantees
to the lifetime of the deserialized string. The default implementation
of `Visitor::visit_borrowed_str` forwards to `visit_str`, meaning
that this change only enables more flexibility.