Speed up parsing with table-driven predicates.
When loading a capture in wrench, startup time is dominated by RON
parsing, and most of that time is spent in simple byte predicates such
as "is this a whitespace byte?"

These predicates currently use `slice::contains` to search through a
list of possible matching bytes. The list lengths range from 4 to 63.

This commit changes these predicates to use a table. Each predicate is
now just an array lookup plus a bit mask. This more than doubles the
parsing speed for the wrench capture I have been looking at.